refactor: 교사 리포트 페이지 로직 분리
